requirement.txt
# Core dependencies
streamlit==1.43.2
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20

# LangChain ecosystem
langchain==0.3.21
langchain-community==0.3.19
langchain-core==0.3.48
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-neo4j==0.4.0
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.15

# Neo4j
neo4j==5.28.1

# AI/ML libraries
openai==1.66.5
huggingface-hub==0.30.1
sentence-transformers==3.4.1
transformers==4.49.0
torch==2.6.0
tokenizers==0.21.1
safetensors==0.5.3
modelscope==1.24.1

# Document processing
pypdf==5.4.0
python-docx==1.2.0
unstructured==0.17.2
filetype==1.2.0

# Data processing
numpy==1.26.4
scikit-learn==1.3.2

# Utilities
colorlog==6.9.0
tqdm==4.67.1
PyYAML==6.0.2
regex==2024.11.6

# Development and testing
pytest==8.3.5
pytest-asyncio==0.26.0
